Rename: js폴더에 pages폴더 추가
