+++ js/pages/common.js
