+++ js/pages/searchResult.js
