+++ js/pages/writePost.js
