+++ js/pages/writePost.ts
